+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Saborter Changelog
 
+## v2.0.0
+
+### Breaking Changes
+
+- Removed the `isError` static method from the `Aborter` class.
+- Removed the static field `errorName` from the `Aborter` class.
+- The `code` and `signal` fields have been removed from `AbortError`.
+- Moved the `dispose` method of the `Aborter` instance into a separate function.
+
+### New Features
+
+- Added utility functions:
+  - catchAbortError
+  - isAbortSignal
+  - rethrowAbortSignal
+  - setTimeoutAsync
+  - throwIfAborted
+  - timeInMilliseconds
+
 ## v1.5.1 (February 11th, 2026)
 
 ### Bug Fixes